First version of some Cornell code ported to run on RA2b's 41x Ntuples made by TheNtupleMaker (https://twiki.cern.ch/twiki/bin/view/CMS/TheNtupleMaker).
Some important features of the original Cornell code are currently missing.
Resurrecting the use of the enums that ultimately specify which type of
collection to use for the various objects and whether or not to turn on various corrections
is important. We may also want to resurrect the class structure of the code.

This analyzer was created with tag v5_1_2 of UserCode/SusySapien/PhysicsTools/TheNtupleMaker.

The variables.txt file used in making the analyzer was created using an ntuple created with tag ngoodmuoc
of UserCode/SusyAnalysis/RA2b/Ntuples/RA2b2011v1 and tag v5_1_1 (I think...) of TheNtupleMaker.
Only the "Events" tree was included.
The file is located here:
/castor/cern.ch/user/s/ssekmen/RA2b2011v1/V00-01-03/QCD_Pt_170to300_TuneZ2_7TeV_pythia6_Spring11-PU_S1_START311_V1G1-v1_AODSIM/QCD_Pt_170to300_TuneZ2_7TeV_pythia6/ntuple_99_4_OhF.root.

I hope we can develop this code in a way that only BasicLoopTools (and maybe BasicLoopCU.cc) has to be modified.
The rest of the files are generated automatically by mkanalyzer.py, and it would be nice
to keep it that way (especially true for BasicLoopCU.h).